Add files via upload
Interpolates data: if you want to fit a data set/model to another data set but x values are different, this will interpolate and create a new dataset 2 with x values corresponding to dataset 1. I used for comparing INS with abINS, where I was trying to fit a blend of abINS spectra derived from DFT phonon calculations to INS data.